pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dprg-archive-agent"
version = "0.1.0"
description = "A tool for searching and chatting with the DPRG archive"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "pinecone-client>=2.0.0",
    "numpy>=1.24.0",
    "rich>=10.0.0",
    "typer>=0.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
namespaces = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "strict" 